python/hosted_agents/hostedagent01_supersimple/pyproject.toml
[project]
name = "hostedagent01-supersimple"
version = "0.1.0"
description = "Minimal hosted agent sample"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-agentserver-agentframework>=1.0.0b17",
    "azure-identity>=1.26.0b2",
    "python-dotenv>=1.0.0",
    "tzdata>=2025.1",
]

[tool.uv]
prerelease = "allow"
